requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
py==1.11.0
PySocks==1.7.1
PyYAML==6.0
requests==2.28.2
retry==0.9.2
selenium==4.8.3
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
wsproto==1.2.0
minio==7.2.3
